+++ gt/tools/auto_rigger/module_biped_finger.py
@@ -885,8 +885,8 @@ if __name__ == "__main__":
     cmds.file(new=True, force=True)
 
     from gt.tools.auto_rigger.rig_framework import RigProject
-    from gt.tools.auto_rigger.rig_module_spine import ModuleSpine
-    from gt.tools.auto_rigger.rig_module_biped_arm import ModuleBipedArmLeft, ModuleBipedArmRight
+    from gt.tools.auto_rigger.module_spine import ModuleSpine
+    from gt.tools.auto_rigger.module_biped_arm import ModuleBipedArmLeft, ModuleBipedArmRight
 
     a_spine = ModuleSpine()
     a_lt_arm = ModuleBipedArmLeft()
