+++ gt/tools/auto_rigger/module_root.py
