+++ gt/tools/auto_rigger/template_biped.py
@@ -2,13 +2,13 @@
 Auto Rigger Project Template for Biped Rigs
 github.com/TrevisanGMW/gt-tools
 """
-from gt.tools.auto_rigger.rig_module_biped_arm import ModuleBipedArmRight, ModuleBipedArmLeft
-from gt.tools.auto_rigger.rig_module_biped_leg import ModuleBipedLegRight, ModuleBipedLegLeft
-from gt.tools.auto_rigger.rig_module_biped_finger import ModuleBipedFingersLeft, ModuleBipedFingersRight
+from gt.tools.auto_rigger.module_biped_arm import ModuleBipedArmRight, ModuleBipedArmLeft
+from gt.tools.auto_rigger.module_biped_leg import ModuleBipedLegRight, ModuleBipedLegLeft
+from gt.tools.auto_rigger.module_biped_finger import ModuleBipedFingersLeft, ModuleBipedFingersRight
 from gt.tools.auto_rigger.rig_framework import RigProject, ModuleGeneric, Proxy
-from gt.tools.auto_rigger.rig_module_spine import ModuleSpine
-from gt.tools.auto_rigger.rig_module_root import ModuleRoot
-from gt.tools.auto_rigger.rig_module_head import ModuleHead
+from gt.tools.auto_rigger.module_spine import ModuleSpine
+from gt.tools.auto_rigger.module_root import ModuleRoot
+from gt.tools.auto_rigger.module_head import ModuleHead
 import maya.cmds as cmds
 import logging
